+++ src/pages/balloonski.tsx
@@ -1,7 +1,8 @@
 import { NextPage } from "next";
 import Head from "next/head";
 import dynamic from "next/dynamic";
-import { ScrollData } from "../themes/Alto/types/scroll";
+import { ScrollData } from "themes/Alto/types/scroll";
+import Renderer from "scenes/Silks/components/Renderer";
 
 const Gotham = dynamic(import("scenes/Gotham"), { ssr: false });
 const Balloonski = dynamic(import("scenes/Gotham/Balloonski"), { ssr: false });
@@ -50,6 +51,7 @@ const LinkTree: NextPage = () => {
         premium
       >
         <Balloonski />
+        {/*<Renderer />*/}
       </Gotham>
     </>
   );
